Fix LLM tests
