+++ apps/ledger-live-mobile/src/families/cardano/Delegations/__tests__/DRepDelegationSelfTransactionInfoDrawer.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { fireEvent, screen, render } from "@tests/test-renderer";
-import DRepDelegationSelfTransactionInfoDrawer from "./DRepDelegationSelfTransactionInfoDrawer";
+import DRepDelegationSelfTransactionInfoDrawer from "../DRepDelegationSelfTransactionInfoDrawer";
 import { CardanoAccount } from "@ledgerhq/live-common/families/cardano/types";
 import { getAccountBridge } from "@ledgerhq/live-common/bridge/index";
 import { NavigatorName, ScreenName } from "~/const";
