+++ apps/ledger-live-mobile/src/families/cardano/UndelegationFlow/__tests__/UndelegationFlow.integ.test.tsx
@@ -6,14 +6,14 @@ import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { component as UndelegationFlow } from "../index";
 import CardanoDelegations from "../../Delegations";
 import { server } from "@tests/server";
-import { handlers } from "../../__tests__/handlers";
 import BigNumber from "bignumber.js";
 import { QueryClientProvider, QueryClient } from "@tanstack/react-query";
-import { ScreenName, NavigatorName } from "~/const";
+import { NavigatorName } from "~/const";
 import { getCardanoAccountFixture } from "@ledgerhq/coin-cardano/fixtures/accounts";
 import { NavigatorScreenParams } from "@react-navigation/native";
 import { CardanoUndelegationFlowParamList } from "../types";
 import * as useBridgeTransaction from "@ledgerhq/live-common/bridge/useBridgeTransaction";
+import { http, HttpResponse } from "msw";
 
 let mockRewardsValue = new BigNumber("0");
 let mockDepositValue = "2000000";
@@ -120,6 +120,47 @@ const INITIAL_STATE = {
 };
 
 describe("UndelegationFlow Integration", () => {
+  const mockPools = [
+    {
+      poolId: "a314a18528d00c5fbd067ecb4a212cf2f307c83d2c08f44a11ebebf6",
+      name: "Ledger by Figment 1",
+      ticker: "LBF1",
+      website: "https://www.ledger.com/coin/staking/cardano",
+      cost: "170.0",
+      margin: "6",
+      pledge: "9.82",
+      liveStake: "40.22",
+      retiredEpoch: 618,
+    },
+    {
+      poolId: "4a9c9902c9538da900b10b716d5d1b214487455fdb06028b32ffa180",
+      name: "Ledger by Figment 2",
+      ticker: "LBF2",
+      website: "https://www.ledger.com/coin/staking/cardano",
+      cost: "170.0",
+      margin: "6",
+      pledge: "9.82",
+      liveStake: "91.69",
+      retiredEpoch: 618,
+    },
+  ];
+
+  const handlers = [
+    http.get("*/v1/pool/list", () => {
+      return HttpResponse.json({
+        pageNo: 1,
+        limit: 10,
+        count: mockPools.length,
+        pools: mockPools,
+      });
+    }),
+    http.get("*/v1/pool/detail", () => {
+      return HttpResponse.json({
+        pools: [mockPools[0]],
+      });
+    }),
+  ];
+
   beforeEach(() => {
     server.use(...handlers);
   });
@@ -147,7 +188,7 @@ describe("UndelegationFlow Integration", () => {
     const stopDelegationBtn = await screen.findByTestId("delegation-undelegate-action");
     await user.press(stopDelegationBtn);
 
-    await screen.findByTestId(ScreenName.CardanoUndelegationSummary);
+    await screen.findByTestId("Cardano-Undelegation-Summary");
 
     expect(screen.getByTestId("undelegation-message")).toBeVisible();
     expect(screen.getByTestId("delegation-undelegate-continue")).toBeEnabled();
